application_manager/requirements.txt
# always required
numpy
blatann

# BLE nordic device
pyserial


# pyqt (pyside) gui toolkit
pyside6
pyqtgraph

# kivy gui toolkit
kivy
kivy_garden.graph